extracted models
