+++ backend/src/models/time.go
@@ -1,0 +1,7 @@
+package models
+
+// Define a struct that represents a time in a Google calendar event
+type Time struct {
+	DateTime string `json:"dateTime"`
+	TimeZone string `json:"timeZone"`
+}
